test: refactor routing tests to reuse fixture
