+++ specs/fixtures/basic/i18n-legacy.config.ts
