+++ specs/fixtures/helpers/layer-path-match-page/nuxt.config.ts
@@ -1,0 +1,1 @@
+export default defineNuxtConfig({})
