--- specs/fixtures/routing/app.vue
@@ -1,11 +1,0 @@
-<template>
-  <NuxtLayout>
-    <NuxtPage />
-  </NuxtLayout>
-</template>
-
-<style>
-section {
-  margin: 1rem 0;
-}
-</style>
