--- specs/fixtures/routing/i18n.config.ts
@@ -1,6 +1,0 @@
-// @ts-ignore
-export default defineI18nConfig(() => {
-  return {
-    legacy: false
-  }
-})
