--- specs/fixtures/routing/nuxt.config.ts
@@ -1,9 +1,0 @@
-export default defineNuxtConfig({
-  // devtools: { enabled: true },
-  modules: ['@nuxtjs/i18n'],
-
-  i18n: {
-    locales: ['en', 'ja'],
-    detectBrowserLanguage: false
-  }
-})
