--- specs/fixtures/routing/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "nuxt3-test-basic",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxi build",
-    "dev": "nuxi dev",
-    "generate": "nuxi generate",
-    "preview": "nuxi preview"
-  },
-  "devDependencies": {
-    "@nuxtjs/i18n": "latest",
-    "nuxt": "latest"
-  }
-}
