--- specs/fixtures/routing/pages/about.vue
@@ -1,3 +1,0 @@
-<template>
-  <div>About</div>
-</template>
