--- specs/fixtures/routing/pages/as a test.vue
@@ -1,3 +1,0 @@
-<template>
-  <div>As a test</div>
-</template>
