--- specs/fixtures/routing/pages/category/[id].vue
@@ -1,5 +1,0 @@
-<script lang="ts" setup></script>
-
-<template>
-  <div>Id</div>
-</template>
