--- specs/fixtures/routing/pages/count/[id].vue
@@ -1,3 +1,0 @@
-<template>
-  <div>Id</div>
-</template>
