--- specs/fixtures/routing/pages/path/[param].vue
@@ -1,3 +1,0 @@
-<template>
-  <div>Testing</div>
-</template>
