+++ specs/routing/legacy.spec.ts
@@ -5,13 +5,16 @@ import { setup } from '../utils'
 import { localeLocationTests, localeRouteTests, switchLocalePathTests } from './routing-tests'
 
 await setup({
-  rootDir: fileURLToPath(new URL(`../fixtures/routing`, import.meta.url)),
+  rootDir: fileURLToPath(new URL(`../fixtures/basic`, import.meta.url)),
   browser: true,
   // overrides
   nuxtConfig: {
+    extends: [fileURLToPath(new URL(`../fixtures/helpers/layer-path-match-page`, import.meta.url))],
     i18n: {
       vueI18n: 'i18n-legacy.config.ts',
       customRoutes: 'config',
+      locales: ['en', 'ja'],
+      defaultLocale: '',
       pages: {
         // 'categories/[id]': {
         //   en: 'categories/english',
